+++ instrumentation/opentelemetry-instrumentation-system-metrics/src/opentelemetry/instrumentation/system_metrics/__init__.py
@@ -36,6 +36,10 @@ following metrics are configured:
         "system.thread_count": None
         "process.runtime.memory": ["rss", "vms"],
         "process.runtime.cpu.time": ["user", "system"],
+        "process.runtime.gc_count": None,
+        "process.runtime.thread_count": None,
+        "process.runtime.cpu.utilization": None,
+        "process.runtime.context_switches": ["involuntary", "voluntary"],
     }
 
 Usage
@@ -63,6 +67,7 @@ Usage
         "system.network.io": ["transmit", "receive"],
         "process.runtime.memory": ["rss", "vms"],
         "process.runtime.cpu.time": ["user", "system"],
+        "process.runtime.context_switches": ["involuntary", "voluntary"],
     }
     SystemMetricsInstrumentor(config=configuration).instrument()
 
@@ -105,6 +110,9 @@ _DEFAULT_CONFIG = {
     "process.runtime.memory": ["rss", "vms"],
     "process.runtime.cpu.time": ["user", "system"],
     "process.runtime.gc_count": None,
+    "process.runtime.thread_count": None,
+    "process.runtime.cpu.utilization": None,
+    "process.runtime.context_switches": ["involuntary", "voluntary"],
 }
 
 
@@ -150,6 +158,9 @@ class SystemMetricsInstrumentor(BaseInstrumentor):
         self._runtime_memory_labels = self._labels.copy()
         self._runtime_cpu_time_labels = self._labels.copy()
         self._runtime_gc_count_labels = self._labels.copy()
+        self._runtime_thread_count_labels = self._labels.copy()
+        self._runtime_cpu_utilization_labels = self._labels.copy()
+        self._runtime_context_switches_labels = self._labels.copy()
 
     def instrumentation_dependencies(self) -> Collection[str]:
         return _instruments
@@ -345,6 +356,29 @@ class SystemMetricsInstrumentor(BaseInstrumentor):
                 callbacks=[self._get_runtime_gc_count],
                 description=f"Runtime {self._python_implementation} GC count",
                 unit="bytes",
+            )
+
+        if "process.runtime.thread_count" in self._config:
+            self._meter.create_observable_up_down_counter(
+                name=f"process.runtime.{self._python_implementation}.thread_count",
+                callbacks=[self._get_runtime_thread_count],
+                description="Runtime active threads count",
+            )
+
+        if "process.runtime.cpu.utilization" in self._config:
+            self._meter.create_observable_gauge(
+                name=f"process.runtime.{self._python_implementation}.cpu.utilization",
+                callbacks=[self._get_runtime_cpu_utilization],
+                description="Runtime CPU utilization",
+                unit="1",
+            )
+
+        if "process.runtime.context_switches" in self._config:
+            self._meter.create_observable_counter(
+                name=f"process.runtime.{self._python_implementation}.context_switches",
+                callbacks=[self._get_runtime_context_switches],
+                description="Runtime context switches",
+                unit="switches",
             )
 
     def _uninstrument(self, **__):
@@ -646,3 +680,34 @@ class SystemMetricsInstrumentor(BaseInstrumentor):
         for index, count in enumerate(gc.get_count()):
             self._runtime_gc_count_labels["count"] = str(index)
             yield Observation(count, self._runtime_gc_count_labels.copy())
+
+    def _get_runtime_thread_count(
+        self, options: CallbackOptions
+    ) -> Iterable[Observation]:
+        """Observer callback for runtime active thread count"""
+        yield Observation(
+            self._proc.num_threads(), self._runtime_thread_count_labels.copy()
+        )
+
+    def _get_runtime_cpu_utilization(
+        self, options: CallbackOptions
+    ) -> Iterable[Observation]:
+        """Observer callback for runtime CPU utilization"""
+        proc_cpu_percent = self._proc.cpu_percent()
+        yield Observation(
+            proc_cpu_percent,
+            self._runtime_cpu_utilization_labels.copy(),
+        )
+
+    def _get_runtime_context_switches(
+        self, options: CallbackOptions
+    ) -> Iterable[Observation]:
+        """Observer callback for runtime context switches"""
+        ctx_switches = self._proc.num_ctx_switches()
+        for metric in self._config["process.runtime.context_switches"]:
+            if hasattr(ctx_switches, metric):
+                self._runtime_context_switches_labels["type"] = metric
+                yield Observation(
+                    getattr(ctx_switches, metric),
+                    self._runtime_context_switches_labels.copy(),
+                )